tests and hex is now unsigned int

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: amaula <amaula@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/11 19:26:00 by amaula            #+#    #+#             */
-/*   Updated: 2024/05/12 16:34:45 by amaula           ###   ########.fr       */
+/*   Updated: 2024/05/13 12:36:11 by amaula           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,7 +23,7 @@ int	ft_putptr(void *p);
 int	ft_putdec(int d);
 int	ft_putint(int i);
 int	ft_putuint(unsigned int i);
-int	ft_puthex(unsigned long h);
-int	ft_puthex_caps(unsigned long hc);
+int	ft_puthex(unsigned int h);
+int	ft_puthex_caps(unsigned int hc);
 
 #endif
